reference espresso, another javascript mode
do this mode in wrong direction:

javascript is not a decent of java or c, indeed it's ancestor is
scheme, so using cc-mode or java-mode as the start point simple does
not make sense.

Implement a real parser or hack a bunch of regular expression to parse
javascript? apparently, write a parser is more complicated, but it
worth to do that to get a real indexer done.

Also, think it from another way: if just want to check the syntax, you
do not need make a parser, a parser is more quick and precise, but
invoke the JSLint via flymake is also good enough.  Anyway, you need a
evaluator when write/test even read the code, so as you can access the
evaluator, parser as syntax checker is not necessory.

It is easy for a solution without parser integrated to mumamo.  But mix
different language in the same file is really a good idea?

js2-mode's problem:

Why not define custom for face?  Not each term support more than 8
colors, and not every developer hate dark theme.  Your syntax hight is
based on a real parser, it is good, but why set such blocks to show
them!?

Not follow the common indent habit is a way to provoke others to
debate, finally just waste the time of the developer.